+++ app/models/concerns/facebook_format.rb
@@ -1,0 +1,93 @@
+module FacebookFormat
+
+  def text_format id, text
+    { recipient: { id: id },
+      message: { text: text }}
+  end
+
+  def button url, title
+    { type: 'web_url',
+      url: url,
+      title: title,
+      webview_height_ratio: 'tall' }
+  end
+
+  def quick_replies_format id, text, title_text=nil, quick_reply_options=nil
+    { recipient: { id: id },
+      message: {
+        text: title_text,
+        quick_replies: quick_reply_options}}
+  end
+
+  def button_format id, text, title_text=nil, button_options=nil
+    { recipient: { id: id },
+      message: {
+        attachment: {
+          type: "template",
+          payload: {
+            template_type: "button",
+            text: title_text,
+            buttons: button_options }}}}
+  end
+
+  def button_option type, title, payload
+    { type: type,
+      title: title,
+      payload: payload }
+  end
+
+  def phone_option title, phone
+    { type: "phone_number",
+      title: title,
+      payload: phone }
+  end
+
+  def button_link_option url, title, webview_height='tall', share_button='hide'
+    { type: 'web_url',
+      url: url,
+      title: title,
+      webview_height_ratio: webview_height,
+      webview_share_button: share_button }
+  end
+
+  def quick_replies_option title, payload
+    { content_type: "text",
+      title: title,
+      payload: payload }
+  end
+
+  def send_location
+    { content_type: "location" }
+  end
+
+  def generic_elements sender_id, results=nil
+    columns = []
+
+    results.each do |result|
+      r = facebook_response(result)
+      r.text = set_text(r, 'facebook')
+
+      actions = []
+      actions << button(safe_url(r.link_url), I18n.t('button.official'))
+      actions << button(safe_url(google.get_map_link(r.lat, r.lng, r.name, r.street)),I18n.t('button.location'))
+      actions << button(safe_url(google.get_google_search(r.name)),I18n.t('button.related_comment'))
+
+      columns << {
+        title: r.name,
+        subtitle: r.text,
+        image_url: r.image_url,
+        buttons: actions }
+    end
+
+    generic_format = {
+      recipient: { id: sender_id },
+      message: {
+        attachment: {
+          type: 'template',
+          payload: {
+            template_type: 'generic',
+            elements: columns
+          }}}}
+    return generic_format
+  end
+end
