+++ app/services/facebook_bot_service.rb
@@ -1,4 +1,5 @@
 class FacebookBotService < BaseService
+  include FacebookFormat
   include Conversion
 
   API_URL ||= "https://graph.facebook.com/v2.6/me/messages?access_token=#{Settings.facebook.page_access_token}"
@@ -205,65 +206,5 @@ class FacebookBotService < BaseService
       options << button_link_option("https://johnwudevelop.tk/users/#{user.id}", '搜尋設定')
       button_format(id, text, title_text, options)
     end
-  end
-
-  def text_format id, text
-    { recipient: { id: id },
-      message: { text: text }}
-  end
-
-  def button url, title
-    { type: 'web_url',
-      url: url,
-      title: title,
-      webview_height_ratio: 'tall' }
-  end
-
-  def quick_replies_format id, text, title_text=nil, quick_reply_options=nil
-    { recipient: { id: id },
-      message: {
-        text: title_text,
-        quick_replies: quick_reply_options}}
-  end
-
-  def button_format id, text, title_text=nil, button_options=nil
-    { recipient: { id: id },
-      message: {
-        attachment: {
-          type: "template",
-          payload: {
-            template_type: "button",
-            text: title_text,
-            buttons: button_options }}}}
-  end
-
-  def button_option type, title, payload
-    { type: type,
-      title: title,
-      payload: payload }
-  end
-
-  def phone_option title, phone
-    { type: "phone_number",
-      title: title,
-      payload: phone }
-  end
-
-  def button_link_option url, title, webview_height='tall', share_button='hide'
-    { type: 'web_url',
-      url: url,
-      title: title,
-      webview_height_ratio: webview_height,
-      webview_share_button: share_button }
-  end
-
-  def quick_replies_option title, payload
-    { content_type: "text",
-      title: title,
-      payload: payload }
-  end
-
-  def send_location
-    { content_type: "location" }
   end
 end
